refactor(META): fix travis and browserstack config
