+++ .travis.yml
@@ -3,6 +3,8 @@ language: node_js
 node_js:
   - '6.10.1'
 script:
+  - export IP_ADDR=$(ip addr | grep eth -A 4 | grep 'inet ' | awk '{ print $2 }' | sed 's/\/..//')
+  - pnpm run build
   - pnpm run test-ci
 before_install:
   - curl -L https://unpkg.com/@pnpm/self-installer | node
@@ -10,7 +12,7 @@ install:
   - pnpm recursive install --filter '@cycle/*'
 cache:
   directories:
-    - /home/travis/.pnpn-store
+    - /home/travis/.pnpm-store
 branches:
   only:
     - master
@@ -19,3 +21,7 @@ addons:
     username: "cyclebot@nada.email"
     access_key:
       secure: "elm+YRK6UYu3MIluSaucd8JWfkif/SVdw0Dp1exoVUA7NL2vF65V6IOOpVC715lfTnoHbbyTFbqwGpll06QJqYa5sumZScF4yg37tRtvEMhEnynAoxKsnhMyMNKefXlLEOavRaUZRs22LyFwEh4JwWt4PrCeBcueNKSdgFnJPTk="
+env:
+  global:
+    - secure: "QbMjL6r0DGAj5WwtLvfyNxJvu4z0WujIWwd0d3aygjXfCQ4jJ6zC2WKEjwh5l5FbjacXJuyJVLC28de+FnPmlWPrI/Pnp1hlxAChCagpd9jZwuCk/6Hd62mTQeWcSz7zdv60MOOGdWfbfLqjgdGy0D7AwsJ/mmfhEROlqQjA5+M="
+    - secure: "VaXjMocfxI4BQBCwDFpEvahhfLxpWoADPA2rmpCsC0mo6EWdy2MW03Td6bQ0B/dQdJ5EDHJcUyuOVhmbimTr6uYDiAAZVn32bZA6m2wPprUt0ZAImNKW/T2Ehe7+GniFmZP9P7slFCX7ZX12UdodfS4PopAKD1HcshBwEphxxIU="
