+++ http/package.json
@@ -19,15 +19,15 @@
     "build-es6": "tsc --module es6 --outDir ./lib/es6",
     "test-node": "mocha test/node.ts --require ts-node/register --exit",
     "test-browser": "pnpm run start-test-support-server && karma start ; pnpm run stop-test-support-server",
-    "start-test-support-server": "PORT=8070 ts-node test/support/run-server.ts & echo $! > test-support-server.pid",
-    "stop-test-support-server": "if ps -p $(cat test-support-server.pid)> /dev/null; then kill -KILL $(cat test-support-server.pid); fi && rm test-support-server.pid",
+    "start-test-support-server": "PORT=8070 ts-node test/support/run-server.ts &",
+    "stop-test-support-server": "kill $(ps aux | grep cyclejs | grep support | grep server | grep -v grep | awk '{ print $2 }') || true",
     "test": "pnpm run test-browser",
     "test-ci": "CI=true pnpm test"
   },
   "devDependencies": {
     "@cycle/isolate": "3.4",
     "@cycle/run": "4.4.0",
-    "@cycle/rxjs-run": "^8.0.0",
+    "@cycle/rxjs-run": "^8.1.0",
     "@types/body-parser": "1.17.x",
     "@types/cookie-parser": "1.4.x",
     "@types/express": "4.16.x",
@@ -38,7 +38,7 @@
     "deepmerge": "^2.1.1",
     "express": "4.16.3",
     "isarray": "^2.0.4",
-    "karma": "2.0.2",
+    "karma": "^3.0.0",
     "karma-browserstack-launcher": "^1.3.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-firefox-launcher": "^1.1.0",
